feature(CopyIcon): Extract copy icon to separate file for reuse
