+++ src/components/CopyIcon/index.tsx
